% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circ_col_chart.R
\name{circ_col_chart}
\alias{circ_col_chart}
\title{Circular Column Chart}
\usage{
circ_col_chart(
  player_data,
  ring_scale,
  ring_inner,
  ring_middle,
  ring_outer,
  ringlab_inner,
  ringlab_middle,
  ringlab_outer,
  col_pal,
  chart_title,
  chart_subtitle,
  chart_source,
  cc_chart_filename
)
}
\arguments{
\item{player_data}{a data frame with a player column and a data column.}

\item{ring_scale}{a numeric value that represent the value the y axis
is scaled by.}

\item{ring_inner}{a numeric value that represents where the inner ring
label will be placed on the graph.}

\item{ring_middle}{a numeric value that represents where the middle ring
label will be placed on the graph.}

\item{ring_outer}{a numeric value that represents where the outer ring
label will be placed on the graph.}

\item{ringlab_inner}{a char value that represents the label of the
inner ring.}

\item{ringlab_middle}{a char value that represents the label of the
middle ring.}

\item{ringlab_outer}{a char value that represents the label of the
outer ring.}

\item{col_pal}{a color palette whose length is the same as the
number of players in the data set.}

\item{chart_title}{is a char value that represents the title
of the radar chart.}

\item{chart_subtitle}{is a char value that is a written explanation
of the information shown on the chart.}

\item{chart_source}{is a char value that represents information about the
data, and visualzation.}

\item{cc_chart_filename}{is a char value that represents the name
the chart file will be saved under.}
}
\value{
A .png file of the total points scored by an NBA team
in the current season over a certain value.
}
\description{
Create a circular column chart.
}
\examples{
player_data <- data.frame(player = c("DA", "CP3", "CJ", "DB", "MB", "CP"),
                       data = c(978, 942, 797, 1789, 1135, 613))
col_pal <- c("#1D1160", "#813840", "#63727A", "#723010", "#000000", "#F9AD1B")
chart_title <- c("Example Plot")
chart_subtitle <- c("This visualization is an example.")
chart_source <- c("\n\n Source: Here \nLink to Data: here")
cc_chart_filename <- c("example_plot.png")
}
