% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pts_radar.R
\name{pts_radar}
\alias{pts_radar}
\title{Points Radar Chart}
\usage{
pts_radar(team_name, slug, team_cols, pts_value, chart_title, pts_filename)
}
\arguments{
\item{team_name}{a char value which represents an NBA name,
which includes the city.}

\item{slug}{a char value that represents uniquely represents the
url for an NBA team on Basket Reference.}

\item{team_cols}{two or more char values which represent the
team colors for an NBA team}

\item{pts_value}{is a numerical value that will filter out
shooters whose total points are less than this value.}

\item{chart_title}{is a char value that represents the title
of the radar chart.}

\item{pts_filename}{is a char value that represents the name
the chart file will be saved under.}
}
\value{
A .png file of the total points scored by an NBA team
in the current season.
}
\description{
Create a radar plot of total points scored in
the current season by a single NBA team.
}
\examples{
team_name <- "Phoenix Suns"
slug <- "PHO"
team_cols <- c("#1D1160", "#E56020", "#63727A", "#000000", "#F9AD1B")
pts_value <- 500
chart_title <- "Phoenix Suns Total Points for 21-22 Season"
pts_filename <- "suns_pts_radar.png"
pts_radar(
    team_name,
    slug,
    team_cols,
    pts_value,
    chart_title,
    pts_filename)
}
